refactor(scan): put interpreter tests in a folder
Per https://github.com/votingworks/vxsuite/pull/1471/files#r819160425
