+++ libs/ballot-interpreter-vx/src/interpreter/basics.test.ts
@@ -1,9 +1,9 @@
 import { BallotType } from '@votingworks/types';
 import { fail } from 'assert';
-import * as choctaw2020LegalSize from '../test/fixtures/choctaw-county-2020-general-election';
-import * as oaklawn from '../test/fixtures/election-4e31cb17d8-ballot-style-77-precinct-oaklawn-branch-library';
-import { Interpreter } from './interpreter';
-import { DetectQrCodeResult } from './types';
+import * as choctaw2020LegalSize from '../../test/fixtures/choctaw-county-2020-general-election';
+import * as oaklawn from '../../test/fixtures/election-4e31cb17d8-ballot-style-77-precinct-oaklawn-branch-library';
+import { Interpreter } from '.';
+import { DetectQrCodeResult } from '../types';
 
 jest.setTimeout(10000);
