+++ libs/ballot-interpreter-vx/src/interpreter/choctaw_2019.test.ts
@@ -1,5 +1,5 @@
-import * as choctaw2019 from '../test/fixtures/election-98f5203139-choctaw-general-2019';
-import { Interpreter } from './interpreter';
+import * as choctaw2019 from '../../test/fixtures/election-98f5203139-choctaw-general-2019';
+import { Interpreter } from '.';
 
 jest.setTimeout(10000);
