+++ libs/ballot-interpreter-vx/src/interpreter/regression_overvote_on_choctaw_county_p1_05.test.ts
@@ -1,5 +1,5 @@
-import * as choctaw2020LegalSize from '../test/fixtures/choctaw-county-2020-general-election';
-import { Interpreter } from './interpreter';
+import * as choctaw2020LegalSize from '../../test/fixtures/choctaw-county-2020-general-election';
+import { Interpreter } from '.';
 
 jest.setTimeout(10000);
