+++ libs/ballot-interpreter-vx/src/interpreter/yesno_overvotes.test.ts
@@ -1,5 +1,5 @@
-import * as hamilton from '../test/fixtures/election-5c6e578acf-state-of-hamilton-2020';
-import { Interpreter } from './interpreter';
+import * as hamilton from '../../test/fixtures/election-5c6e578acf-state-of-hamilton-2020';
+import { Interpreter } from '.';
 
 jest.setTimeout(10000);
